feat: move connectors to a separate package
