+++ packages/common/src/index.ts
@@ -1,4 +1,3 @@
-export * from './connectors';
 export * from './constants';
 export * from './types';
 export * from './utils';
