+++ packages/common/src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from './address';
 export * from './bytes';
-export * from './device';
 export * from './errors';
 export * from './get-interface';
 export * from './regexps';
