+++ packages/connectors/.eslintignore
@@ -1,0 +1,3 @@
+dist/
+jest.config.js
+jest-base.js
