+++ packages/connectors/package.json
@@ -1,0 +1,109 @@
+{
+  "name": "@flair-sdk/connectors",
+  "version": "2.47.0",
+  "private": false,
+  "author": "flair.dev",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/flair-sdk/typescript.git"
+  },
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org/"
+  },
+  "scripts": {
+    "build": "rollup -c"
+  },
+  "peerDependencies": {
+    "@0xsequence/wagmi-connector": "*",
+    "@ethersproject/providers": "*",
+    "@everipedia/wagmi-magic-connector": "*",
+    "@gnosis.pm/safe-apps-provider": "*",
+    "@gnosis.pm/safe-apps-sdk": "*",
+    "@wagmi/core": "*",
+    "@web3-onboard/common": "*",
+    "@web3-onboard/core": "*",
+    "@web3-onboard/injected-wallets": "*",
+    "@web3-onboard/ledger": "*",
+    "@web3-onboard/portis": "*",
+    "@web3auth/ui": "*",
+    "@web3auth/web3auth-wagmi-connector": "*",
+    "bowser": "*",
+    "ethers": "*",
+    "eventemitter3": "*"
+  },
+  "peerDependenciesMeta": {
+    "@ethersproject/providers": {
+      "optional": true
+    },
+    "@everipedia/wagmi-magic-connector": {
+      "optional": true
+    },
+    "@gnosis.pm/safe-apps-provider": {
+      "optional": true
+    },
+    "@gnosis.pm/safe-apps-sdk": {
+      "optional": true
+    },
+    "@wagmi/core": {
+      "optional": true
+    },
+    "ethers": {
+      "optional": true
+    },
+    "bowser": {
+      "optional": true
+    },
+    "eventemitter3": {
+      "optional": true
+    },
+    "@0xsequence/wagmi-connector": {
+      "optional": true
+    },
+    "@web3-onboard/common": {
+      "optional": true
+    },
+    "@web3-onboard/core": {
+      "optional": true
+    },
+    "@web3-onboard/injected-wallets": {
+      "optional": true
+    },
+    "@web3-onboard/ledger": {
+      "optional": true
+    },
+    "@web3-onboard/portis": {
+      "optional": true
+    },
+    "@web3auth/ui": {
+      "optional": true
+    },
+    "@web3auth/web3auth-wagmi-connector": {
+      "optional": true
+    }
+  },
+  "dependencies": {
+    "@0xsequence/wagmi-connector": "^1.0.1",
+    "@everipedia/wagmi-magic-connector": "^0.6.5",
+    "@flair-sdk/common": "^2.47.0",
+    "@gnosis.pm/safe-apps-provider": "^0.13.1",
+    "@gnosis.pm/safe-apps-sdk": "^7.8.0",
+    "@wagmi/core": "^0.5.7",
+    "@web3-onboard/common": "^2.2.3",
+    "@web3-onboard/core": "^2.10.0",
+    "@web3-onboard/injected-wallets": "^2.3.0",
+    "@web3-onboard/ledger": "^2.3.2",
+    "@web3-onboard/portis": "^2.1.3",
+    "@web3auth/ui": "^3.0.4",
+    "@web3auth/web3auth-wagmi-connector": "^1.0.0",
+    "bowser": "^2.11.0",
+    "ethers": "^5.6.5"
+  },
+  "main": "dist/index.js",
+  "module": "dist/index.es.js",
+  "types": "dist/types.d.ts",
+  "files": [
+    "dist"
+  ]
+}
