+++ packages/connectors/rollup.config.js
@@ -1,0 +1,32 @@
+import commonjs from '@rollup/plugin-commonjs';
+import json from '@rollup/plugin-json';
+import resolve from '@rollup/plugin-node-resolve';
+import typescript from '@rollup/plugin-typescript';
+import dts from 'rollup-plugin-dts';
+import peerDepsExternal from 'rollup-plugin-peer-deps-external';
+
+import packageJson from './package.json';
+
+export default [
+  {
+    input: './src/index.ts',
+    output: [
+      {
+        file: packageJson.main,
+        format: 'cjs',
+        sourcemap: true,
+      },
+      {
+        file: packageJson.module,
+        format: 'esm',
+        sourcemap: true,
+      },
+    ],
+    plugins: [peerDepsExternal(), resolve(), json(), commonjs(), typescript()],
+  },
+  {
+    input: 'src/index.ts',
+    output: [{ file: 'dist/types.d.ts' }],
+    plugins: [dts()],
+  },
+];
