+++ packages/connectors/src/utils/device.ts
